Don't display missing shutter or aperture value
Previously when filling in just one field, e.g. shutter,
the display would look like:

    f/0 - 1/200 s

In case of missing aperture, we should only show the shutter value:

    1/200 s

Created a formatting function to perform the creation of this whole
string with all the conditional logic.

I dropped the use of StringBuilder - according to StackOverflow the
compiler should compile the string concatenation to use of
StringBuilder anyway... so why complicate the code?

- https://stackoverflow.com/questions/18453458/string-builder-vs-string-concatenation

Also dropped the use of TextUtil.SPACE - I don't really see a
point of declaring a constant for " " character.